% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{aggregate_taxa}
\alias{aggregate_taxa}
\title{Organize data in higher taxonomic level}
\usage{
aggregate_taxa(obj, rank)
}
\arguments{
\item{obj}{MetaGenomic object}

\item{rank}{taxonomic level choosen}
}
\description{
Organize data in higher taxonomic level
}
