% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{arrange_taxa}
\alias{arrange_taxa}
\title{Enlarge taxa in MetaGenomic}
\usage{
arrange_taxa(obj, new_taxa)
}
\arguments{
\item{obj}{\code{MetaGenomic} object}

\item{new_taxa}{data.frame with larger taxonomic information}
}
\value{
\code{MetaGenomic} object with enlarged taxonomic information. The new taxa in the data.frame data will be represented by columns of zeros.
}
\description{
Replaces a larger taxonomy data.frame than the present one. All taxa present in obj must also be present in the new taxa_new object. It is used to compare different MetaGenomic object.
}
