% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-class.R
\name{arrange_vertices}
\alias{arrange_vertices}
\title{Arrange the vertex number}
\usage{
arrange_vertices(obj, new_taxonomy)
}
\arguments{
\item{new_taxonomy}{data.frame with taxonomic information about the vertices wanted in new graph.
The dataset must have the same form as taxa in MetaGenomic/MetaGenomicGraph object.
Naturally in new_taxonomy all the old vertices must be present.}
}
\description{
Edit the vertices / taxa of the MetaGenomicGraph object (for example to compare two different networks):
}
