% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-utility.R
\name{community_detection}
\alias{community_detection}
\title{Get Communities of a Signed Weighted Graph}
\usage{
community_detection(graph)
}
\description{
Adaptation in R of the algorithm for detention of weighted communities with sign of a graph developed by Sergio Gomez.
https://deim.urv.cat/~sergio.gomez/radatools.php
}
