% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-utility.R
\name{coord.positive.link}
\alias{coord.positive.link}
\title{Signed-Weighted Graph Layout}
\usage{
coord.positive.link(graph, seed = 123)
}
\description{
It elaborates the coordinates for the representation of
the vertices of the graph considering only the links with a positive sign.
}
