% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-class.R
\name{default_decoration}
\alias{default_decoration}
\title{Add graphical decorations}
\usage{
default_decoration(obj)
}
\description{
Add graphical descriptions to network:
\itemize{
\item Vertex size proportional to mean clr abundances over samples.
\item Link color could be red (+) or blue (-) respect the weights sign.
}
}
