% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-utility.R
\name{density_link_threshold}
\alias{density_link_threshold}
\title{Edge Density Threshold}
\usage{
density_link_threshold(weights, edge.density = 0.05)
}
\arguments{
\item{weights}{correlation matrix}

\item{edge.density}{threshold value}
}
\description{
Get the adjacency matrix considering only the highest absolute value weights
up to the desired link density.
}
