% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-utility.R
\name{get_gl_weights}
\alias{get_gl_weights}
\title{Spiec-Easi GLASSO weights extraction}
\usage{
get_gl_weights(res.gl, Refit = TRUE)
}
\description{
Extract the weight matrix of the GLASSO criterion
in the Spiec-Easi algorithm.
}
