% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-utility.R
\name{get_mb_weights}
\alias{get_mb_weights}
\title{Spiec-Easi MB weights extraction}
\usage{
get_mb_weights(res.mb)
}
\arguments{
\item{res.mb}{Spiec-Easi algorithm results with mb criteria}
}
\description{
Extract the weight matrix of the Meinshausen and Bühlmann criterion
in the Spiec-Easi algorithm.
}
