% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{is.MetaGenomic}
\alias{is.MetaGenomic}
\title{Check if object belong to MetaGenomic Class}
\usage{
is.MetaGenomic(obj)
}
\arguments{
\item{object}{object to be check}
}
\description{
Check if object belong to MetaGenomic Class
}
