% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{MetaGenomic-class}
\alias{MetaGenomic-class}
\alias{MetaGenomic}
\title{S4 class to manage metagenomic data}
\usage{
MetaGenomic(data, meta, taxa)
}
\description{
S4 class to manage metagenomic data
}
\section{Slots}{

\describe{
\item{\code{data}}{table of data}

\item{\code{meta}}{sample features}

\item{\code{taxa}}{taxonomic information}
}}

