% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-class.R
\name{MetaGenomicGraph-class}
\alias{MetaGenomicGraph-class}
\alias{MetaGenomicGraph}
\title{S4 class to manage metagenomic data}
\usage{
MetaGenomicGraph(MetaGenomic, Adjacency)
}
\description{
S4 class to manage metagenomic data
}
\section{Slots}{

\describe{
\item{\code{data}}{table of data}

\item{\code{meta}}{sample features}

\item{\code{taxa}}{taxonomic information}

\item{\code{netw}}{Metagenomic igraph network}

\item{\code{comm}}{graph communities}
}}

