% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-class.R
\name{MetaGenomicGraph}
\alias{MetaGenomicGraph}
\title{MeteGenomicGraph Constructor}
\usage{
MetaGenomicGraph(MetaGenomic, Adjacency)
}
\arguments{
\item{MetaGenomic}{Metagenomic Class Object.}

\item{Corr}{Adjacency Matrix.}
}
\description{
MeteGenomicGraph Constructor
}
