% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{pairwise}
\alias{pairwise}
\title{Apply function to columns/elements pairwise
Literally copied from the package AkselA/R-ymse: Ymse!!!}
\usage{
pairwise(x, FUN, ..., comm = FALSE)
}
\arguments{
\item{x}{a matrix or data.frame}

\item{FUN}{any function that takes two vectors as input and returns a single
value}

\item{...}{further arguments passed to FUN}

\item{comm}{logical; is FUN commutative? If true, only the lower
triangle, including the diagonal, is computed}
}
\value{
An \eqn{n\times n}{n×n} square matrix with \eqn{n} the number of columns
of \code{x}.
}
\description{
Pairwise application of a function to the columns of a matrix/data.frame or
elements of a list
}
\examples{
dtf <- data.frame(aa=c(1, 1, 2, 2, 3, 2, 4),
                  bb=c(1, 1, 2, 3, 3, 3, 4),
                  cc=c(3, 3, 2, 1, 1, 1, 1),
                  dd=c(1, 2, 2, 2, 1, 1, 2))

# Root Mean Square Deviation
pairwise(dtf, function(x, y) sqrt(mean((x-y)^2)))

# using with cor.test() to accompany cor()
pv <- pairwise(dtf, function(x, y) cor.test(x, y)$p.val)
pvn <- 6^(1.1-pv)-5
pvn[pvn<1] <- 1

set_mar(1, 1, 1, 1)
plot(0, xlim=c(0.5, 4.5), ylim=c(0.5, 4.5), cex=0, ann=FALSE, xaxt="n", yaxt="n")
text(rep(1:4, 4), rep(4:1, each=4), t(round(cor(dtf), 2)), cex=pvn,
  col=c("black", "darkgrey")[(pv>0.1)+1])

\url{https://rdrr.io/github/AkselA/R-ymse/src/R/pairwise.R}

}
\seealso{
\code{\link{similarity}} for a few more examples
}
