% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-utility.R
\name{remove.small.communities}
\alias{remove.small.communities}
\title{remove.small.communities}
\usage{
remove.small.communities(obj, vertex.num.threshold, keep.discarded)
}
\description{
remove.small.communities
}
\author{
Alessandro Fuschi
}
