% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{sample_selection}
\alias{sample_selection}
\title{Smart sample selection}
\usage{
sample_selection(obj, condition = "AND", ...)
}
\arguments{
\item{obj}{\code{MetaGenomic} object}

\item{condition}{in progress}

\item{...}{in progress}
}
\value{
\code{MetaGenomic} object with select samples according to the features choose from meta table.
}
\description{
Allows you to choose samples based on the information stored in the meta table
}
