% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{[,MetaGenomic-method}
\alias{[,MetaGenomic-method}
\title{Method extensions to subset MetaGenomic objects.}
\usage{
\S4method{[}{MetaGenomic}(x, i, j, ..., drop = TRUE)
}
\description{
Method extensions to subset MetaGenomic objects.
}
