% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomicGraph-class.R
\name{[,MetaGenomicGraph,ANY,ANY,ANY-method}
\alias{[,MetaGenomicGraph,ANY,ANY,ANY-method}
\title{Method extensions to subset MetaGenomicGraph objects.}
\usage{
\S4method{[}{MetaGenomicGraph,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Method extensions to subset MetaGenomicGraph objects.
}
