% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{taxa_filtering}
\alias{taxa_filtering}
\title{Smart taxa filtering}
\usage{
taxa_filtering(
  obj,
  condition,
  prevalence,
  detection,
  median_non_zero,
  min_count,
  relative_abundance,
  perc_count_lost,
  keep_discarded
)
}
\arguments{
\item{obj}{MetaGenomic object}

\item{condition}{in progress}

\item{prevalence}{in progress}

\item{detection}{in progress}

\item{median_non_zero}{in progress}

\item{relative_abundance}{in progress}

\item{data}{in progress}
}
\value{
\code{MetaGenomic} object with select taxa.
}
\description{
Filter the rarest taxa based on the conditions imposed in the parameters.
}
