% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaGenomic-class.R
\name{taxa_name}
\alias{taxa_name}
\title{Get taxa name}
\usage{
taxa_name(obj, rank)
}
\arguments{
\item{obj}{MetaGenomic object}

\item{rank}{taxonomic level choosen (if not set, the finest taxonomic rank is assumed)}
}
\description{
Get taxa name
}
