% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{tril}
\alias{tril}
\title{Lower triangular matrix}
\usage{
tril(X, diag = FALSE)
}
\arguments{
\item{X}{matrix or data.frame}

\item{diag}{includes or not the diagonal}
}
\description{
Returns the lower triangular values of a square matrix as one-dimensional array.
}
