% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{triu}
\alias{triu}
\title{Upper triangular matrix}
\usage{
triu(X, diag = FALSE)
}
\arguments{
\item{X}{matrix or data.frame}

\item{diag}{includes or not the diagonal}
}
\description{
Returns the upper triangular values of a square matrix as one-dimensional array.
}
