% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{unroll}
\alias{unroll}
\title{Expresses a square numeric matrix as a list of position indices and values}
\usage{
unroll(X, diag = FALSE, sparse = TRUE, symmetric = TRUE)
}
\arguments{
\item{X}{matrix or data.frame}

\item{diag}{includes or not include the diagonal}

\item{sparse}{removes zero values}

\item{symmetric}{if true select only the upper triangular values}
}
\description{
Converts a square matrix or numeric data.frame to a the list of all
values described with respect to the position of the indices.
If are present present rows and cols names the information are adds like additional columns.
}
